+++ R/src_postgres.R
@@ -40,7 +40,7 @@ src_postgres <- function(dbname = "test",
 
   # add plpgsql function for docdb_update(), which should only raise
   # an error if the plpgsql extension is not installed in dbname
-  if (inherits(try(
+  out <- try(
     # credits: Joao Haas, https://stackoverflow.com/a/65093455
     DBI::dbExecute(conn = con, statement = 'CREATE OR REPLACE FUNCTION
      jsonb_merge_patch("target" jsonb, "patch" jsonb)
@@ -62,7 +62,11 @@ src_postgres <- function(dbname = "test",
      WHERE jsonb_typeof("pval") != \'null\' OR "pval" ISNULL;
     END;
    $$ LANGUAGE plpgsql;
-  '), silent = TRUE), "try-error")) {
+  '), silent = TRUE)
+
+  if (inherits(out, "try-error") &&
+      !grepl("tuple concurrently updated", out)) {
+
     stop("PostgreSQL does not support plpgsql in table '", dbname,
          "', but this is needed for nodbi::docdb_update()")
   }
@@ -73,7 +77,6 @@ src_postgres <- function(dbname = "test",
     f = closeNodbiConnections,
     onexit = TRUE
   )
-  # message("nodbi src_postgres reg.finalizer registered")
 
   # return standard nodbi structure
   structure(list(con = con,